refactor(cli): split workspace command owners
