+++ framework/core/src/python/kungfu/cli/commands/_workspace/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: Apache-2.0
